Move grid view mode to its own /grid route (#1551)
Split the climb list page into separate /list and /grid routes using a
Next.js route group (climbs) for shared layout. This code-splits grid
components (ClimbCard, GridClimbItem) into a separate chunk, reducing
initial JS for list-mode users. View toggle buttons are now Link
components with route-level prefetching, and the user's preference is
persisted in IndexedDB to generate correct URLs throughout the app.

Key changes:
- New useViewModePreference hook with module-level cache for sync reads
- URL constructors accept optional viewMode parameter ('list' | 'grid')
- ClimbsList split into shared shell + list-view-content + grid-view-content
- Shared SSR helper (fetch-climb-page-data.ts) for both page routes
- All 9 client navigation call sites pass view mode preference to URLs
- Updated tests for new viewMode parameter and grid path handling
