+++ packages/web/app/[board_name]/[layout_id]/[size_id]/[set_ids]/[angle]/(climbs)/grid/page.tsx
@@ -1,0 +1,63 @@
+import React from 'react';
+
+import { notFound, permanentRedirect } from 'next/navigation';
+import { BoardRouteParametersWithUuid, SearchRequestPagination } from '@/app/lib/types';
+import { constructClimbListWithSlugs } from '@/app/lib/url-utils';
+import { parseRouteParams } from '@/app/lib/url-utils.server';
+import { getBoardDetailsForBoard } from '@/app/lib/board-utils';
+import BoardPageGridView from '@/app/components/board-page/board-page-grid-view';
+import { fetchClimbPageData } from '@/app/components/board-page/fetch-climb-page-data';
+
+export default async function GridPage(props: {
+  params: Promise<BoardRouteParametersWithUuid>;
+  searchParams: Promise<SearchRequestPagination>;
+}) {
+  const searchParams = await props.searchParams;
+  const params = await props.params;
+  const { parsedParams, isNumericFormat } = await parseRouteParams(params);
+
+  // Redirect old numeric URLs to new slug format (always redirect to /grid)
+  if (isNumericFormat) {
+    const boardDetails = getBoardDetailsForBoard(parsedParams);
+
+    if (boardDetails.layout_name && boardDetails.size_name && boardDetails.set_names) {
+      const newUrl = constructClimbListWithSlugs(
+        boardDetails.board_name,
+        boardDetails.layout_name,
+        boardDetails.size_name,
+        boardDetails.size_description,
+        boardDetails.set_names,
+        parsedParams.angle,
+        'grid',
+      );
+
+      // Preserve search parameters
+      const searchString = new URLSearchParams(
+        Object.entries(searchParams).reduce(
+          (acc, [key, value]) => {
+            if (value !== undefined) {
+              acc[key] = String(value);
+            }
+            return acc;
+          },
+          {} as Record<string, string>,
+        ),
+      ).toString();
+      const finalUrl = searchString ? `${newUrl}?${searchString}` : newUrl;
+
+      permanentRedirect(finalUrl);
+    }
+  }
+
+  const data = await fetchClimbPageData(parsedParams, searchParams);
+  if (!data) return notFound();
+
+  return (
+    <>
+      {data.preloadUrl && (
+        <link rel="preload" as="image" href={data.preloadUrl} fetchPriority="high" />
+      )}
+      <BoardPageGridView {...parsedParams} boardDetails={data.boardDetails} initialClimbs={data.climbs} />
+    </>
+  );
+}
