+++ packages/web/app/[board_name]/[layout_id]/[size_id]/[set_ids]/[angle]/(climbs)/layout-client.tsx
