+++ packages/web/app/[board_name]/[layout_id]/[size_id]/[set_ids]/[angle]/(climbs)/layout.tsx
@@ -13,7 +13,7 @@ interface LayoutProps {
   params: Promise<BoardRouteParameters>;
 }
 
-export default async function ListLayout(props: PropsWithChildren<LayoutProps>) {
+export default async function ClimbsLayout(props: PropsWithChildren<LayoutProps>) {
   const params = await props.params;
 
   const { children } = props;
