+++ packages/web/app/b/[board_slug]/[angle]/(climbs)/layout.tsx
@@ -3,13 +3,13 @@ import { PropsWithChildren } from 'react';
 import { notFound } from 'next/navigation';
 import { resolveBoardBySlug, boardToRouteParams } from '@/app/lib/board-slug-utils';
 import { getBoardDetailsForBoard } from '@/app/lib/board-utils';
-import ListLayoutClient from '@/app/[board_name]/[layout_id]/[size_id]/[set_ids]/[angle]/list/layout-client';
+import ListLayoutClient from '@/app/[board_name]/[layout_id]/[size_id]/[set_ids]/[angle]/(climbs)/layout-client';
 
 interface LayoutProps {
   params: Promise<{ board_slug: string; angle: string }>;
 }
 
-export default async function BoardSlugListLayout(props: PropsWithChildren<LayoutProps>) {
+export default async function BoardSlugClimbsLayout(props: PropsWithChildren<LayoutProps>) {
   const params = await props.params;
   const { children } = props;
